pwm: Support PWM with sigma delta HW
to provide PWM frequencies of >312kHz. This feature can for example be
used to provide a step up converter to convert the supply voltage to
a higher voltage. The following circuit provides an example how to do
this:

       +-----+       Diode
 VCC +-+  L  +------+-|>|-+--+ OUT
       +-----+      |     |
                 T  |    +++
       +-----+   .--/    | |
GPIO +-+  R  +--(--> )   |C|
       +-----+   `--\    | |
                    |    +++
                    |     |
                    +--+--+
                      GND

The output voltage could for example be used to drive string lights
which require an input voltage of 32V.
To calculate the right L, [1] might be helpful.

The following table describes which resolutions and frequencies for
the PWM are supported with the ESP8266;

CPU clock->|   80MHz  |  160MHz  |
Resolution |PWM frequency (in Hz)|
-----------|----------|----------|
     8 bit |   312,500|   625,000|
     7 bit |   625,000| 1,250,000|
     6 bit | 1,250,000| 2,500,000|
     5 bit | 2,500,000| 5,000,000|
     4 bit | 5,000,000|10,000,000|
     3 bit |10,000,000|20,000,000|
     2 bit |20,000,000|40,000,000|

[1] http://schmidt-walter-schaltnetzteile.de/smps_e/aww_smps_e.html

Signed-off-by: Timo Wischer <twischer@freenet.de>