settings: Assert on changed Settings size
When modifying the Settings structure it is quite helpful to get a
feedback in case the size of the Settings structure has unexpectedly
changed. A modification changing the size of Settings would not
be backward compatible and even stop the ESP8266 from booting.
Therefore it is quite helpful to get this feedback as early as
possible.

Signed-off-by: Timo Wischer <twischer@freenet.de>